+++ src/de/hhu/bsinfo/dxram/boot/AbstractBootComponent.java
@@ -23,7 +23,7 @@ public abstract class AbstractBootComponent extends AbstractDXRAMComponent {
 	 * @param p_priorityInit     Default init priority for this component
 	 * @param p_priorityShutdown Default shutdown priority for this component
 	 */
-	public AbstractBootComponent(final int p_priorityInit, final int p_priorityShutdown) {
+	public AbstractBootComponent(final short p_priorityInit, final short p_priorityShutdown) {
 		super(p_priorityInit, p_priorityShutdown);
 	}
